moving recurring expenses files to specific package
