+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/recurringexpenses/RecurringExpensesScreen.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.screens
+package br.com.rbrthmn.ui.financialcompanion.screens.recurringexpenses
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.LinearOutSlowInEasing
